Clean up, add separate config.hh and window_handle.hh
