+++ src/phantasm-renderer/backend/d3d12/Swapchain.hh
@@ -4,7 +4,7 @@
 
 #include <typed-geometry/types/size.hh>
 
-#include <phantasm-renderer/backend/types.hh>
+#include <phantasm-renderer/backend/config.hh>
 
 #include "common/d3d12_sanitized.hh"
 #include "common/shared_com_ptr.hh"
