+++ src/phantasm-renderer/backend/types.hh
@@ -2,12 +2,6 @@
 
 #include <cstdint>
 
-struct SDL_Window;
-typedef struct HWND__* HWND;
-typedef struct _XDisplay Display;
-typedef unsigned long XID;
-typedef XID Window;
-
 namespace pr::backend
 {
 namespace handle
@@ -104,127 +98,6 @@ enum class queue_type : uint8_t
     graphics,
     copy,
     compute
-};
-
-enum class adapter_preference : uint8_t
-{
-    highest_vram,
-    first,
-    integrated,
-    highest_feature_level,
-    explicit_index
-};
-
-enum class validation_level : uint8_t
-{
-    off,
-
-    // D3D12: Whether to enable debug layers, requires installed D3D12 SDK
-    // Vulkan: Whether to enable validation, requires installed LunarG SDK
-    on,
-
-    // D3D12: Whether to additionally enable GPU based validation (slow)
-    //
-    // Vulkan: Whether to additionally enable LunarG GPU-assisted validation
-    //          Slow, and requires a reserved descriptor set. If your device
-    //          only has 8 (max shader args * 2), like an IGP, this could fail
-    //
-    // Extended validation for both APIs can prevent diagnostic tools like
-    // Renderdoc and NSight from working properly (PIX will work though)
-    on_extended,
-
-    // D3D12: Whether to additionally enable DRED (Device Removed Extended Data)
-    //          with automatic breadcrumbs and pagefault recovery (very slow)
-    //          see: https://docs.microsoft.com/en-us/windows/win32/direct3d12/use-dred
-    //
-    // Vulkan: No additional effect
-    on_extended_dred
-};
-
-enum class present_mode : uint8_t
-{
-    allow_tearing,
-    synced
-};
-
-/// Special features that are backend-specific, ignored if not applicable
-PR_DEFINE_BIT_FLAGS(native_feature_flags, uint32_t){
-    none = 0,
-    /// Vulkan: Enables VK_LAYER_LUNARG_api_dump (prints all API calls to stdout)
-    vk_api_dump = 0x0001,
-};
-}
-
-struct backend_config
-{
-    /// whether to enable API-level validations
-    validation_level validation = validation_level::off;
-
-    /// the swapchain presentation mode (note: synced does not mean refreshrate-limited)
-    /// Also note: under some circumstances, synced might force a refreshrate limit (Nvidia optimus + windowed on d3d12, etc.)
-    present_mode present_mode = present_mode::synced;
-
-    /// the strategy for choosing a physical GPU
-    adapter_preference adapter_preference = adapter_preference::highest_vram;
-    unsigned explicit_adapter_index = unsigned(-1);
-
-    /// native features to enable
-    native_feature_flags_t native_features = native_feature_flags::none;
-
-    /// whether to enable DXR / VK raytracing features if available
-    bool enable_raytracing = true;
-
-    /// amount of backbuffers to create
-    unsigned num_backbuffers = 3;
-
-    /// amount of threads to accomodate
-    /// backend calls must only be made from <= [num_threads] unique OS threads
-    unsigned num_threads = 1;
-
-    /// resource limits
-    unsigned max_num_resources = 2048;
-    unsigned max_num_pipeline_states = 1024;
-    unsigned max_num_cbvs = 2048;
-    unsigned max_num_srvs = 2048;
-    unsigned max_num_uavs = 2048;
-    unsigned max_num_samplers = 1024;
-    unsigned max_num_accel_structs = 2048;
-    unsigned max_num_raytrace_pipeline_states = 256;
-
-    /// command list allocator size (total = #threads * #allocs/thread * #lists/alloc)
-    unsigned num_cmdlist_allocators_per_thread = 5;
-    unsigned num_cmdlists_per_allocator = 5;
-};
-
-/// opaque window handle
-struct native_window_handle
-{
-    enum wh_type : uint8_t
-    {
-        wh_sdl,
-        wh_win32_hwnd,
-        wh_xlib
-    };
-
-    wh_type type;
-
-    union {
-        ::SDL_Window* sdl_handle;
-        ::HWND win32_hwnd;
-        struct
-        {
-            ::Window window;
-            ::Display* display;
-        } xlib_handles;
-    } value;
-
-    native_window_handle(::HWND hwnd) : type(wh_win32_hwnd) { value.win32_hwnd = hwnd; }
-    native_window_handle(::SDL_Window* sdl_window) : type(wh_sdl) { value.sdl_handle = sdl_window; }
-    native_window_handle(::Window xlib_win, ::Display* xlib_display) : type(wh_xlib)
-    {
-        value.xlib_handles.window = xlib_win;
-        value.xlib_handles.display = xlib_display;
-    }
 };
 
 // Maps to
